requirements.txt
django==1.9.5
django-bootstrap-form==3.2
django-widget-tweaks==1.4.1
humanize==0.5.1
ipdb==0.9.0
ipython==4.1.2
mysql-python==1.2.5
Pillow==3.1.1
pyasn1==0.1.9
pyOpenSSL==16.0.0
python-dateutil==2.5.1
python-mimeparse==1.5.1
python-openid==2.2.5
pytz==2016.3
requests==2.9.1
requests-oauthlib==0.6.1
simplejson==3.8.2
six==1.10.0
tablib==0.11.2
Unidecode==0.4.19